Update the ABI 6 build contract and string lengths
Require libmustache 0.6 and C++17 while preserving pkg-config compiler flags, and configure Windows static consumers appropriately.

Use Zend-provided lengths for PHP strings crossing the C++ boundary without changing ABI 6 ownership representations.

+++ mustache_template.cpp
@@ -58,7 +58,7 @@ PHP_METHOD(MustacheTemplate, __construct)
   try {
     // Custom parameters
     char * template_str = NULL;
-    long template_len = 0;
+    size_t template_len = 0;
 
     // Check parameters
     zval * _this_zval = NULL;
@@ -114,4 +114,3 @@ PHP_METHOD(MustacheTemplate, __toString)
   }
 }
 /* }}} MustacheTemplate::__toString */
-
